Reorganize integration tests.
