+++ test/database/util.go
@@ -1,0 +1,15 @@
+package database
+
+import (
+	"github.com/NebulousLabs/skynet-accounts/database"
+)
+
+// DBTestCredentials sets the environment variables to what we have defined in Makefile.
+func DBTestCredentials() database.DBCredentials {
+	return database.DBCredentials{
+		User:     "admin",
+		Password: "aO4tV5tC1oU3oQ7u",
+		Host:     "localhost",
+		Port:     "17017",
+	}
+}
